Refactor opengl methods from polyset-gl.cc into renderer.cc
CMakeLists.txt: Remove polyset-gl.cc from build
openscad.pro: Remove polyset-gl.cc from build
src/CGALRenderer.cc: Call new parent render_edges and render_surface methods
src/OpenCSGRenderer.cc: Pass renderer to OpenCSGPrim to remove static call
  for Renderer::render_surface.  Add OpenCSGPrim private reference variable
  for renderer.  Pass this renderer when creating new OpenCSGPrim object.
src/polyset-gl.cc: Delete file
src/polyset.h: Remove system-gl.h and renderer.h.  Add method to get internal
  polygon member variable.  Remove render_surface and render_edges methods.
src/renderer.cc: Make get_csgmode non-static member and const.
  Move polyset-gl.cc static functions. Remove static render_surface and
  render_edges methods.
src/renderer.h: Add Geometry.h header. Convert static methods to virtual.
