+++ CMakeLists.txt
@@ -501,7 +501,6 @@ set(NOCGAL_SOURCES
   src/export_svg.cc
   src/LibraryInfo.cc
   src/polyset.cc
-  src/polyset-gl.cc
   src/polyset-utils.cc
   src/GeometryUtils.cc)
