+++ openscad.pro
@@ -444,7 +444,6 @@ SOURCES += \
            src/polyset-utils.cc \
            src/GeometryUtils.cc \
            src/polyset.cc \
-           src/polyset-gl.cc \
            src/csgops.cc \
            src/transform.cc \
            src/color.cc \
