+++ src/polyset.h
@@ -1,10 +1,8 @@
 #pragma once
 
 #include "Geometry.h"
-#include "system-gl.h"
 #include "linalg.h"
 #include "GeometryUtils.h"
-#include "renderer.h"
 #include "Polygon2d.h"
 #include <vector>
 #include <string>
@@ -20,6 +18,8 @@ public:
 	PolySet(unsigned int dim, boost::tribool convex = unknown);
 	PolySet(const Polygon2d &origin);
 	~PolySet();
+
+	const Polygon2d &getPolygon() const { return polygon; }
 
 	size_t memsize() const override;
 	BoundingBox getBoundingBox() const override;
@@ -39,9 +39,6 @@ public:
 	void insert_vertex(const Vector3d &v);
 	void insert_vertex(const Vector3f &v);
 	void append(const PolySet &ps);
-
-	void render_surface(Renderer::csgmode_e csgmode, const Transform3d &m, GLint *shaderinfo = nullptr) const;
-	void render_edges(Renderer::csgmode_e csgmode) const;
 
 	void transform(const Transform3d &mat);
 	void resize(const Vector3d &newsize, const Eigen::Matrix<bool,3,1> &autosize);
