+++ src/renderer.h
@@ -5,6 +5,7 @@
 #include "memory.h"
 #include "colormap.h"
 #include "enums.h"
+#include "Geometry.h"
 
 #ifdef _MSC_VER // NULL
 #include <cstdlib>
@@ -48,9 +49,9 @@ public:
 	virtual Color4f setColor(ColorMode colormode, const float color[4], GLint *shaderinfo = nullptr) const;
 	virtual void setColorScheme(const ColorScheme &cs);
 
-	static csgmode_e get_csgmode(const bool highlight_mode, const bool background_mode, const OpenSCADOperator type=OpenSCADOperator::UNION);
-	static void render_surface(shared_ptr<const class Geometry> geom, csgmode_e csgmode, const Transform3d &m, GLint *shaderinfo = nullptr);
-	static void render_edges(shared_ptr<const Geometry> geom, csgmode_e csgmode);
+	virtual csgmode_e get_csgmode(const bool highlight_mode, const bool background_mode, const OpenSCADOperator type=OpenSCADOperator::UNION) const;
+	virtual void render_surface(shared_ptr<const class Geometry> geom, csgmode_e csgmode, const Transform3d &m, GLint *shaderinfo = nullptr) const;
+	virtual void render_edges(shared_ptr<const Geometry> geom, csgmode_e csgmode) const;
 
 protected:
 	std::map<ColorMode,Color4f> colormap;
